ENH: Restructuring.
